+++ tests/test_self_checks.py
@@ -2,12 +2,16 @@ import textwrap
 
 import django.core.checks
 
-from extra_checks.checks.model_field_checks import (
+from extra_checks.check_id import CheckId
+from extra_checks.checks import (
+    CheckConfig,
     CheckFieldFileUploadTo,
     CheckFieldForeignKeyIndex,
+    check_extra_checks_health,
 )
-from extra_checks.checks.self_checks import CheckConfig, check_extra_checks_health
+from extra_checks.checks.base_checks import BaseCheck
 from extra_checks.registry import Registry
+from extra_checks.utils import collect_subclasses
 
 
 def test_empty_config(registry, settings):
@@ -52,3 +56,13 @@ def test_error_formatting(registry, settings):
             """
         ).strip()
     )
+
+
+def test_unique_check_ids():
+    used_checks = [
+        c.Id for c in collect_subclasses(BaseCheck.__subclasses__()) if hasattr(c, "Id")
+    ]
+    unused = CheckId._value2member_map_.keys() - set(used_checks)
+    assert not unused, "Not all CheckIds used."
+    dups = {c.value for c in used_checks if used_checks.count(c) > 1}
+    assert not dups, "CheckIds must be unique per Check."